Fixed pathes in bash scripts

+++ app/scripts/backup.sh
@@ -8,8 +8,13 @@
 # @package    deployment
 # @author     André Lademann <vergissberlin@googlemail.com>
 
-# 1. Backup settings
-cp -u /usr/src/app/app/node-red/flows_cred.json /data/backup/node-red/flows_cred.json
-
-# 2. Backup flow
+# 1. Backup flow
 cp -u /usr/src/app/app/node-red/flows.json /data/backup/node-red/flows.json
+
+# 2. Backup credentials
+if [ -e "/usr/src/app/app/node-red/flows_cred.json" ]
+  then
+    cp -u /usr/src/app/app/node-red/flows_cred.json /data/backup/node-red/flows_cred.json
+  else
+    echo "No credentials found."
+fi
