+++ app/scripts/recover.sh
@@ -8,8 +8,13 @@
 # @package    deployment
 # @author     André Lademann <vergissberlin@googlemail.com>
 
-# 1. Recover settings
-cp -f /data/backup/node-red/flows_cred.json /usr/src/app/app/node-red/flows_cred.json
-
-# 2. Recover flow
+# 1. Recover flow
 cp -f /data/backup/node-red/flows.json /usr/src/app/app/node-red/flows.json
+
+# 2. Recover credentials
+if [ -e "/data/backup/node-red/flows_cred.json" ]
+  then
+    cp -f /data/backup/node-red/flows_cred.json /usr/src/app/app/node-red/flows_cred.json
+  else
+    echo "No credentials found."
+fi
